Create a common utility for downloading. This allows emulaotr and universal maker to share some code.
