+++ src/downloadUtils.ts
@@ -1,4 +1,5 @@
 import { URL } from "url";
+import * as crypto from "crypto";
 import * as fs from "fs-extra";
 import * as ProgressBar from "progress";
 import * as tmp from "tmp";
@@ -12,7 +13,7 @@ import { FirebaseError } from "./error";
  * @param remoteUrl URL to download.
  * @param auth Whether to include an access token in the download request. Defaults to false.
  */
-export async function downloadToTmp(remoteUrl: string, auth: boolean = false): Promise<string> {
+export async function downloadToTmp(remoteUrl: string, auth = false): Promise<string> {
   const u = new URL(remoteUrl);
   const c = new Client({ urlPrefix: u.origin, auth });
   const tmpfile = tmp.fileSync();
@@ -46,4 +47,47 @@ export async function downloadToTmp(remoteUrl: string, auth: boolean = false): P
   });
 
   return tmpfile.name;
+}
+
+/**
+ * Checks whether the file at `filepath` has the expected size.
+ */
+export function validateSize(filepath: string, expectedSize: number): Promise<void> {
+  return new Promise((resolve, reject) => {
+    const stat = fs.statSync(filepath);
+    return stat.size === expectedSize
+      ? resolve()
+      : reject(
+          new FirebaseError(
+            `download failed, expected ${expectedSize} bytes but got ${stat.size}`,
+            { exit: 1 },
+          ),
+        );
+  });
+}
+
+/**
+ * Checks whether the file at `filepath` has the expected checksum.
+ */
+export function validateChecksum(
+  filepath: string,
+  expectedChecksum: string,
+  algorithm: "md5" | "sha256" = "md5",
+): Promise<void> {
+  return new Promise((resolve, reject) => {
+    const hash = crypto.createHash(algorithm);
+    const stream = fs.createReadStream(filepath);
+    stream.on("data", (data: any) => hash.update(data));
+    stream.on("end", () => {
+      const checksum = hash.digest("hex");
+      return checksum === expectedChecksum
+        ? resolve()
+        : reject(
+            new FirebaseError(
+              `download failed, expected checksum ${expectedChecksum} but got ${checksum}`,
+              { exit: 1 },
+            ),
+          );
+    });
+  });
 }
